+++ Common/src/main/java/fuzs/universalenchants/mixin/accessor/EnchantmentAccessor.java
@@ -1,5 +1,6 @@
 package fuzs.universalenchants.mixin.accessor;
 
+import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import org.spongepowered.asm.mixin.Mixin;
@@ -12,4 +13,8 @@ public interface EnchantmentAccessor {
     @Accessor
     @Mutable
     void setCategory(EnchantmentCategory category);
+
+    @Accessor
+    @Mutable
+    void setSlots(EquipmentSlot[] slots);
 }